+++ src/assets/scripts/client/aircraft/AircraftTypeDefinitionModel.js
@@ -1,10 +1,11 @@
 import BaseModel from '../base/BaseModel';
 import { INVALID_NUMBER } from '../constants/globalConstants';
 import { isEmptyObject } from '../utilities/validatorUtilities';
+import { AIRPORT_CONSTANTS } from '../constants/airportConstants';
 
 // TODO: abstract these to an appropriate constants file
 const HEAVY_LETTER = 'H';
-const SUPER_LETTER = 'U';
+const SUPER_LETTER = 'J';
 
 /**
  * Provides a definition for a specific type of aircraft.
@@ -49,7 +50,7 @@ export default class AircraftTypeDefinitionModel extends BaseModel {
         this.icao = '';
 
         /**
-         * Icao identifier that includes a weightclass
+         * Icao identifier that includes a weightClass
          * designation when `Heavy` or `Super`
          *
          * @property icaoWithWeightClass
@@ -68,11 +69,11 @@ export default class AircraftTypeDefinitionModel extends BaseModel {
         this.engines = null;
 
         /**
-         * @property weightclass
+         * @property weightClass
          * @type {string}
          * @default ''
          */
-        this.weightclass = '';
+        this.weightClass = '';
 
         /**
          * @property category
@@ -152,7 +153,7 @@ export default class AircraftTypeDefinitionModel extends BaseModel {
         this.name = aircraftTypeDefinition.name;
         this.icao = aircraftTypeDefinition.icao.toLowerCase();
         this.engines = aircraftTypeDefinition.engines;
-        this.weightclass = aircraftTypeDefinition.weightclass;
+        this.weightClass = aircraftTypeDefinition.weightClass;
         this.category = aircraftTypeDefinition.category;
         this.ceiling = aircraftTypeDefinition.ceiling;
         this.rate = aircraftTypeDefinition.rate;
@@ -174,7 +175,7 @@ export default class AircraftTypeDefinitionModel extends BaseModel {
         this.icao = '';
         this.icaoWithWeightClass = '';
         this.engines = null;
-        this.weightclass = '';
+        this.weightClass = '';
         this.category = null;
         this.ceiling = INVALID_NUMBER;
         this.rate = null;
@@ -194,13 +195,17 @@ export default class AircraftTypeDefinitionModel extends BaseModel {
     _buildTypeForStripView() {
         let aircraftIcao = `${this.icao}/L`;
 
-        switch (this.weightclass) {
+        switch (this.weightClass) {
             case SUPER_LETTER:
+                aircraftIcao = `${SUPER_LETTER}/${this.icao}/L`;
+
+                break;
             case HEAVY_LETTER:
                 aircraftIcao = `${HEAVY_LETTER}/${this.icao}/L`;
 
                 break;
             default:
+
                 break;
         }
 
@@ -233,6 +238,40 @@ export default class AircraftTypeDefinitionModel extends BaseModel {
      * @returns {Boolean}
      */
     isHeavyOrSuper() {
-        return this.weightclass === HEAVY_LETTER || this.weightclass === SUPER_LETTER;
+        return this.weightClass === HEAVY_LETTER || this.weightClass === SUPER_LETTER;
+    }
+
+    /**
+     * Returns the minimal distance that is required to a previous aircraft before another aircraft is allowed to use the runway.
+     *
+     * @for AircraftTypeDefinitionModel
+     * @method calculateSameRunwaySeparationDistanceInFeet
+     * @param previousTypeModel {AircraftTypeDefinitionModel} the aircraft type that used the runway before us.
+     * @returns {number} distance in feet
+     */
+    calculateSameRunwaySeparationDistanceInFeet(previousTypeModel) {
+        if (previousTypeModel.isSameRunwaySeparationCatThree()) {
+            return AIRPORT_CONSTANTS.SRS_REDUCED_MINIMA_FEET.CAT3;
+        }
+
+        switch (this.category.srs) {
+            case 1:
+                return AIRPORT_CONSTANTS.SRS_REDUCED_MINIMA_FEET.CAT1;
+            case 2:
+                return AIRPORT_CONSTANTS.SRS_REDUCED_MINIMA_FEET.CAT2;
+            default:
+                return AIRPORT_CONSTANTS.SRS_REDUCED_MINIMA_FEET.CAT3;
+        }
+    }
+
+    /**
+     * Returns true if srs cat 3 is required
+     *
+     * @for AircraftTypeDefinitionModel
+     * @method isSameRunwaySeparationCatThree
+     * @returns true if srs cat 3 is required
+     */
+    isSameRunwaySeparationCatThree() {
+        return typeof this.category.srs === 'undefined' || this.category.srs === 3;
     }
 }