+++ test/aircraft/Pilot/Pilot.spec.js
@@ -1,7 +1,6 @@
 import ava from 'ava';
 import sinon from 'sinon';
 import _isArray from 'lodash/isArray';
-import _isEqual from 'lodash/isEqual';
 import _isObject from 'lodash/isObject';
 import AircraftModel from '../../../src/assets/scripts/client/aircraft/AircraftModel';
 import ModeController from '../../../src/assets/scripts/client/aircraft/ModeControl/ModeController';
@@ -28,9 +27,10 @@ const runwayModelMock = airportModelFixture.getRunway(runwayNameMock);
 const approachTypeMock = 'ils';
 
 const validRouteStringMock = 'DAG.KEPEC3.KLAS07R';
-const complexRouteString = 'COWBY..BIKKR..DAG.KEPEC3.KLAS';
-const amendRouteString = 'HITME..HOLDM..BIKKR';
+const complexRouteString = 'COWBY..BIKKR..DAG.KEPEC3.KLAS01L';
+const amendRouteString = 'DAG..HOLDM..PRINO';
 const invalidRouteString = 'A..B.C.D';
+const invalidAmendRouteString = 'A..B..C';
 const sidIdMock = 'COWBY6';
 const waypointNameMock = 'SUNST';
 const holdParametersMock = {
@@ -94,7 +94,7 @@ ava('.applyArrivalProcedure() returns an error when passed an invalid routeStrin
     const pilot = createPilotFixture();
     const result = pilot.applyArrivalProcedure('~!@#$%', airportNameMock);
 
-    t.true(_isEqual(result, expectedResult));
+    t.deepEqual(result, expectedResult);
 });
 
 ava('.applyArrivalProcedure() returns an error when passed an invalid procedure name', (t) => {
@@ -103,7 +103,7 @@ ava('.applyArrivalProcedure() returns an error when passed an invalid procedure 
     const pilot = createPilotFixture();
     const result = pilot.applyArrivalProcedure(invalidRouteStringMock, airportNameMock);
 
-    t.true(_isEqual(result, expectedResult));
+    t.deepEqual(result, expectedResult);
 });
 
 ava('.applyArrivalProcedure() returns an error when passed a procedure with an invaild entry', (t) => {
@@ -112,7 +112,7 @@ ava('.applyArrivalProcedure() returns an error when passed a procedure with an i
     const pilot = createPilotFixture();
     const result = pilot.applyArrivalProcedure(invalidRouteStringMock, airportNameMock);
 
-    t.true(_isEqual(result, expectedResult));
+    t.deepEqual(result, expectedResult);
 });
 
 ava('.applyArrivalProcedure() returns a success message after success', (t) => {
@@ -134,52 +134,47 @@ ava('.applyArrivalProcedure() calls #_fms.replaceArrivalProcedure() with the cor
     t.true(replaceArrivalProcedureSpy.calledWithExactly(validRouteStringMock));
 });
 
-ava.skip('.applyDepartureProcedure() returns an error when passed an invalid sidId', (t) => {
-    const expectedResult = [false, 'SID name not understood'];
+ava('.applyDepartureProcedure() returns an error when passed an invalid sidId', (t) => {
+    const procedureName = '~!@#$%';
+    const expectedResult = [false, 'unknown procedure "~!@#$%"'];
     const pilot = new Pilot(createFmsDepartureFixture(), createModeControllerFixture(), createNavigationLibraryFixture());
-    const result = pilot.applyDepartureProcedure('~!@#$%', airportIcaoMock);
+    const result = pilot.applyDepartureProcedure(procedureName, airportIcaoMock);
 
-    t.true(_isEqual(result, expectedResult));
+    t.deepEqual(result, expectedResult);
     t.false(pilot.hasDepartureClearance);
 });
 
-ava.skip('.applyDepartureProcedure() returns an error when passed an invalid runway', (t) => {
-    const expectedResult = [false, 'unsure if we can accept that procedure; we don\'t have a runway assignment'];
+ava('.applyDepartureProcedure() returns an error when passed an invalid runway', (t) => {
+    const routeString = 'EDDF30R.COWBY6.GUP';
+    const expectedResult = [false, 'requested route of "EDDF30R.COWBY6.GUP" is invalid'];
     const pilot = new Pilot(createFmsDepartureFixture(), createModeControllerFixture(), createNavigationLibraryFixture());
-    const result = pilot.applyDepartureProcedure(sidIdMock, null, airportIcaoMock);
+    const result = pilot.applyDepartureProcedure(routeString, airportIcaoMock);
 
-    t.true(_isEqual(result, expectedResult));
+    t.deepEqual(result, expectedResult);
     t.false(pilot.hasDepartureClearance);
 });
 
-ava.skip('.applyDepartureProcedure() returns an error when passed a runway incompatable for the route', (t) => {
-    const expectedResult = [false, 'unable, the COWBOY SIX departure not valid from Runway ~!@#$%'];
-    const invalidRunwayModelMock = {
-        name: '~!@#$%'
-    };
+ava('.applyDepartureProcedure() should NOT change mcp modes', (t) => {
     const pilot = new Pilot(createFmsDepartureFixture(), createModeControllerFixture(), createNavigationLibraryFixture());
-    const result = pilot.applyDepartureProcedure(sidIdMock, invalidRunwayModelMock, airportIcaoMock);
+    const mcp = pilot._mcp;
+    const expectedAltitudeMode = mcp.altitudeMode;
+    const expectedSpeedMode = mcp.speedMode;
 
-    t.true(_isEqual(result, expectedResult));
-    t.false(pilot.hasDepartureClearance);
-});
-
-ava.skip('.applyDepartureProcedure() should set mcp altitude and speed modes to `VNAV`', (t) => {
-    const pilot = new Pilot(createFmsDepartureFixture(), createModeControllerFixture(), createNavigationLibraryFixture());
     pilot.applyDepartureProcedure(sidIdMock, airportIcaoMock);
 
-    t.true(pilot._mcp.altitudeMode === 'VNAV');
-    t.true(pilot._mcp.speedMode === 'VNAV');
+    // workaround: t.true(pilot._mcp..altitudeMode) causes out of memory crash
+    t.true(mcp.altitudeMode === expectedAltitudeMode);
+    t.true(mcp.speedMode === expectedSpeedMode);
 });
 
-ava.skip('.applyDepartureProcedure() returns a success message after success', (t) => {
+ava('.applyDepartureProcedure() returns a success message after success', (t) => {
     const pilot = new Pilot(createFmsDepartureFixture(), createModeControllerFixture(), createNavigationLibraryFixture());
     const result = pilot.applyDepartureProcedure(sidIdMock, airportIcaoMock);
 
     t.true(_isArray(result));
     t.true(result[0]);
     t.true(result[1].log === 'cleared to destination via the COWBY6 departure, then as filed');
-    t.true(result[1].say === 'cleared to destination via the Cowboy Six departure, then as filed');
+    t.true(result[1].say === 'cleared to destination via the COWBOY SIX departure, then as filed');
 });
 
 ava('.replaceFlightPlanWithNewRoute() returns an error when passed an invalid route', (t) => {
@@ -193,7 +188,7 @@ ava('.replaceFlightPlanWithNewRoute() returns an error when passed an invalid ro
     const pilot = createPilotFixture();
     const result = pilot.replaceFlightPlanWithNewRoute('a..b.c.d');
 
-    t.true(_isEqual(result, expectedResult));
+    t.deepEqual(result, expectedResult);
 });
 
 ava('.replaceFlightPlanWithNewRoute() removes an existing route and replaces it with a new one', (t) => {
@@ -217,46 +212,97 @@ ava('.replaceFlightPlanWithNewRoute() returns a success message when finished su
     const pilot = createPilotFixture();
     const result = pilot.replaceFlightPlanWithNewRoute('COWBY..BIKKR');
 
-    t.true(_isEqual(result, expectedResult));
+    t.deepEqual(result, expectedResult);
 });
 
-ava.skip('.applyPartialRouteAmendment() returns an error with passed an invalid routeString', (t) => {
+ava('.applyPartialRouteAmendment() returns an error with passed an invalid routeString', (t) => {
     const expectedResult = [false, 'requested route of "A..B.C.D" is invalid'];
     const pilot = buildPilotWithComplexRoute();
     const result = pilot.applyPartialRouteAmendment(invalidRouteString);
 
-    t.true(_isEqual(result, expectedResult));
+    t.deepEqual(result, expectedResult);
 });
 
-ava.skip('.applyPartialRouteAmendment() returns an error with passed a routeString without a shared waypoint', (t) => {
-    const expectedResult = [false, 'requested route of "HITME..HOLDM" is invalid, it must contain a Waypoint in the current route'];
+ava('.applyPartialRouteAmendment() returns an error with passed a routeString without a shared waypoint', (t) => {
+    const expectedResult = [false, 'routes do not have continuity!'];
     const pilot = buildPilotWithComplexRoute();
     const result = pilot.applyPartialRouteAmendment('HITME..HOLDM');
 
-    t.true(_isEqual(result, expectedResult));
+    t.deepEqual(result, expectedResult);
 });
 
-ava.skip('.applyPartialRouteAmendment() returns to the correct flightPhase after a hold', (t) => {
-    const pilot = buildPilotWithComplexRoute();
-    pilot._fms.setFlightPhase('HOLD');
-
-    pilot.applyPartialRouteAmendment(amendRouteString);
-
-    t.true(pilot._fms.currentPhase === 'CRUISE');
-});
-
-ava.skip('.applyPartialRouteAmendment() returns a success message when complete', (t) => {
+ava('.applyPartialRouteAmendment() returns a success message when complete', (t) => {
     const expectedResult = [
         true,
         {
-            log: 'rerouting to: HITME..HOLDM..BIKKR..DAG.KEPEC3.KLAS',
+            log: 'rerouting to: DAG HOLDM PRINO',
             say: 'rerouting as requested'
         }
     ];
     const pilot = buildPilotWithComplexRoute();
     const result = pilot.applyPartialRouteAmendment(amendRouteString);
 
-    t.true(_isEqual(result, expectedResult));
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.applyPartialRouteAmendment() calls #_fms.applyPartialRouteAmendment()', (t) => {
+    const pilot = buildPilotWithComplexRoute();
+    const fmsApplyPartialRouteAmendmentSpy = sinon.spy(pilot._fms, 'applyPartialRouteAmendment');
+    const expectedResult = [
+        true,
+        {
+            log: 'rerouting to: DAG HOLDM PRINO',
+            say: 'rerouting as requested'
+        }
+    ];
+    const result = pilot.applyPartialRouteAmendment(amendRouteString);
+
+    t.true(fmsApplyPartialRouteAmendmentSpy.calledWithExactly(amendRouteString));
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.applyPartialRouteAmendment() does not grant departure clearance when the route amendment fails', (t) => {
+    const pilot = buildPilotWithComplexRoute();
+
+    pilot.hasDepartureClearance = false;
+    t.false(pilot.hasDepartureClearance);
+
+    const expectedResult = [false, `requested route of "${invalidAmendRouteString}" is invalid`];
+    const result = pilot.applyPartialRouteAmendment(invalidAmendRouteString);
+
+    t.deepEqual(result, expectedResult);
+    t.false(pilot.hasDepartureClearance);
+});
+
+ava('.applyPartialRouteAmendment() grants departure clearance when the route amendment succeeds', (t) => {
+    const pilot = buildPilotWithComplexRoute();
+
+    pilot.hasDepartureClearance = false;
+    t.false(pilot.hasDepartureClearance);
+
+    const expectedResult = [true,
+        {
+            log: 'rerouting to: DAG HOLDM PRINO',
+            say: 'rerouting as requested'
+        }
+    ];
+    const result = pilot.applyPartialRouteAmendment(amendRouteString);
+
+    t.deepEqual(result, expectedResult);
+
+    // workaround: t.true(pilot.hasDepartureClearance) causes out of memory crash
+    const departureClearance = pilot.hasDepartureClearance;
+    t.true(departureClearance);
+});
+
+ava('.applyPartialRouteAmendment() calls .exitHold()', (t) => {
+    const pilot = buildPilotWithComplexRoute();
+    const exitHoldSpy = sinon.spy(pilot, 'exitHold');
+
+    pilot.initiateHoldingPattern('MISEN', holdParametersMock);
+    pilot.applyPartialRouteAmendment(amendRouteString);
+
+    t.true(exitHoldSpy.calledWithExactly());
 });
 
 ava('.cancelApproachClearance() returns early if #hasApproachClearance is false', (t) => {
@@ -418,9 +464,9 @@ ava('.conductInstrumentApproach() returns failure message when no runway is prov
 ava('.conductInstrumentApproach() returns failure message when assigned altitude is lower than minimum glideslope intercept altitude', (t) => {
     const expectedResult = [false, {
         log: 'unable ILS 19L, our assigned altitude is below the minimum glideslope ' +
-        'intercept altitude, request climb to 2800',
+        'intercept altitude, request climb to 3400',
         say: 'unable ILS one niner left, our assigned altitude is below the minimum ' +
-        'glideslope intercept altitude, request climb to two thousand eight hundred'
+        'glideslope intercept altitude, request climb to three thousand four hundred'
     }];
     const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK, createNavigationLibraryFixture());